split index
